backend/pyproject.toml
[project]
name = "q-backend"
version = "0.1.0"
description = "Q queue management backend (FastAPI + SQLAlchemy + WebSockets)"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    # pydantic 2.10+ ships pre-built 3.13 wheels for pydantic-core.
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",
    "PyJWT[cryptography]==2.10.1",
    "bcrypt==4.2.1",
    "python-multipart==0.0.20",
    "qrcode[pil]==8.0",
    "httpx==0.28.1",
    "websockets==14.1",
    "greenlet==3.1.1",
    "slowapi==0.1.9",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "aiosqlite==0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-q -ra"
